+++ pkg/morph/event/reputation/put_notary.go
@@ -3,6 +3,7 @@ package reputation
 import (
 	"fmt"
 
+	"github.com/nspcc-dev/neo-go/pkg/smartcontract/scparser"
 	"github.com/nspcc-dev/neofs-node/pkg/morph/event"
 )
 
@@ -24,12 +25,6 @@ func (p *Put) setValue(v []byte) error {
 	return p.value.Unmarshal(v)
 }
 
-var fieldSetters = []func(*Put, []byte) error{
-	// order on stack is reversed
-	(*Put).setValue,
-	(*Put).setPeerID,
-}
-
 const (
 	// PutNotaryEvent is method name for reputation put operations
 	// in `Reputation` contract. Is used as identifier for notary
@@ -39,36 +34,39 @@ const (
 
 // ParsePutNotary from NotaryEvent into reputation event structure.
 func ParsePutNotary(ne event.NotaryEvent) (event.Event, error) {
-	var ev Put
+	const putArgCnt = 3
+	var ev = new(Put)
 
-	fieldNum := 0
+	args := ne.Params()
+	if len(args) != putArgCnt {
+		return nil, event.WrongNumberOfParameters(putArgCnt, len(args))
+	}
 
-	for _, op := range ne.Params() {
-		switch fieldNum {
-		case 0, 1:
-			data, err := event.BytesFromOpcode(op)
-			if err != nil {
-				return nil, err
-			}
+	epoch, err := scparser.GetInt64FromInstr(args[0].Instruction)
+	if err != nil {
+		return nil, fmt.Errorf("epoch: %w", err)
+	}
+	ev.setEpoch(uint64(epoch))
 
-			err = fieldSetters[fieldNum](&ev, data)
-			if err != nil {
-				return nil, fmt.Errorf("can't parse field num %d: %w", fieldNum, err)
-			}
-		case 2:
-			n, err := event.IntFromOpcode(op)
-			if err != nil {
-				return nil, err
-			}
+	peerID, err := scparser.GetBytesFromInstr(args[1].Instruction)
+	if err != nil {
+		return nil, fmt.Errorf("peer ID: %w", err)
+	}
+	err = ev.setPeerID(peerID)
+	if err != nil {
+		return nil, fmt.Errorf("peer ID: %w", err)
+	}
 
-			ev.setEpoch(uint64(n))
-		default:
-			return nil, event.UnexpectedArgNumErr(PutNotaryEvent)
-		}
-		fieldNum++
+	value, err := scparser.GetBytesFromInstr(args[2].Instruction)
+	if err != nil {
+		return nil, fmt.Errorf("value: %w", err)
+	}
+	err = ev.setValue(value)
+	if err != nil {
+		return nil, fmt.Errorf("value: %w", err)
 	}
 
 	ev.notaryRequest = ne.Raw()
 
-	return ev, nil
+	return *ev, nil
 }